refact:decimal e renome getId

+++ ApiCar/Models/FuelRecord.cs
@@ -7,7 +7,7 @@ public class FuelRecord
     [Key] public int Id { get; set; }
     [Required] public int CarId { get; set; }
     [Required] public int Quantity { get; set; }
-    [Required] public int Price { get; set; }
+    [Required] public decimal Price { get; set; }
     [Required] public DateTime Date { get; set; }
     [Required] public int Mileage { get; set; }
     public virtual Car? Car { get; set; }
